Error handling float to int

+++ src/edu/ggc/it/gpacalc/GPACalcActivity.java
@@ -67,7 +67,8 @@ public class GPACalcActivity extends Activity
                 //Error checking for empty Hours text field
                 if (!stringHours.isEmpty())
                 {
-                    hours = Integer.parseInt(stringHours);
+                    
+                    hours = (int) Float.parseFloat(stringHours);
                 }
                 else 
                 {
